# api/generated/swagger-api/swagger.yaml
...
swagger: "2.0"
info:
  version: "v1"
  title: "BringAlong Server REST API"
schemes:
- "https"
paths:
...
  /marketplace/v1/bringdemand/{id}:
    get:
      operationId: "MarketplaceRestService_getBringDemand_get"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/BringDemandEto"
    delete:
      operationId: "MarketplaceRestService_deleteBringDemand_delete"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        default:
          description: "successful operation"
...
definitions:
...
  BringDemandEto:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      modificationCounter:
        type: "integer"
        format: "int32"
      item:
        type: "string"
      dropHint:
        type: "string"
      deadline:
        type: "string"
        format: "date-time"
      estimatedAmount:
        type: "integer"
        format: "int32"
      currency:
        $ref: "#/definitions/Currency"
      acceptedByUser:
        type: "integer"
        format: "int64"
      userId:
        type: "integer"
        format: "int64"
...
  Currency:
    type: "object"
    properties:
      currencyCode:
        type: "string"
      defaultFractionDigits:
        type: "integer"
        format: "int32"
      numericCode:
        type: "integer"
        format: "int32"
      displayName:
        type: "string"
      symbol:
        type: "string"
      numericCodeAsString:
        type: "string"
...